fix: cast drift_pct expression to numeric before ROUND() in vector_status()

+++ src/api/diagnostics.rs
@@ -1829,7 +1829,7 @@ pub(super) fn vector_status() -> TableIterator<
                    CASE \
                      WHEN c.reltuples > 0 \
                      THEN ROUND( \
-                       (COALESCE(s.rows_changed_since_last_reindex, 0)::float / c.reltuples) * 100.0, \
+                       ((COALESCE(s.rows_changed_since_last_reindex, 0)::float / c.reltuples) * 100.0)::numeric, \
                        2) \
                      ELSE NULL \
                    END AS drift_pct \
